Test pydbus on GLib 2.40 (Ubuntu 14.04) too.

+++ tests/run.sh
@@ -15,14 +15,20 @@ rm "$ADDRESS_FILE" "$PID_FILE"
 
 PYTHON=${1:-python}
 
-"$PYTHON" -m pydbus.examples.notifications_server &
-NOTIF_PID=$!
-trap 'kill -TERM $PID $NOTIF_PID' EXIT
+if [ "$2" != "dontpublish" ]
+then
+	"$PYTHON" -m pydbus.examples.notifications_server &
+	NOTIF_PID=$!
+	trap 'kill -TERM $PID $NOTIF_PID' EXIT
+fi
 
 sleep 1
 
 "$PYTHON" -m pydbus.tests.context
 "$PYTHON" -m pydbus.tests.identifier
-"$PYTHON" -m pydbus.tests.publish
-"$PYTHON" -m pydbus.tests.publish_properties
-"$PYTHON" -m pydbus.tests.publish_multiface
+if [ "$2" != "dontpublish" ]
+then
+	"$PYTHON" -m pydbus.tests.publish
+	"$PYTHON" -m pydbus.tests.publish_properties
+	"$PYTHON" -m pydbus.tests.publish_multiface
+fi
